Calculadora en básica en C

+++ dividir.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include "dividir.h"
 
-float dividir(int valor1, int valor2){
+float dividir(float valor1, float valor2){
     return valor1 / valor2;
 }
